Move inspector utils to utils package
